requirements.txt
currencyconverter
requests
tensorflow-gpu